Upgrade Flask to 1.1.4
Pin requirements to 1.1 or greater, but less than 2. Flask 1.1 is the
last planned v1 release before 2.0.0, so take patch versions of 1.1 for
security but not 2.0 which we're not ready for yet.

Take new versions of utils and content-loader which have been updated
for Flask 1.1

Also remove the pin for `itsdangerous` - Flask now controls the
supported versions. See https://github.com/pallets/flask/blob/1.1.4/setup.py#L58